backend/accounts: separate tx proposal from note proposal
When making a tx in the send screen, a tx proposal is created when
form fields are changed, to store the tx proposal to be signed and
sent, and also to compute the tx fees that the frontend should
display. This endpoint can have a long delay, e.g. in Ethereum it does
network requests.

The note was also part of the tx proposal, but that lead to a needless
delay and a display of "calculating fees..." when typing a note.

This commit makes a separate tx note proposal, as it can always be
processed instantly.