frontend/feetargets: add custom fee target
The custom fee target is added in expert mode.

When an expert user wants to set custom fee, it will be helpful to see
what the actual rates are for the four priorities, so those are also
provided to the frontend.

The actual UI and where to place them can be improved in a later
commit.