frontend: support custom fee input
Make input field editable if costum expert setting is active
and feetarget is custom. Display total fee below the input
field for now.
The feetarget take more space now as the target does show
sat/vb info, that is why it will break to its own line on
medium sized screens and only is shown in columns on large
screens.