frontend/web/preact: remove unused typings-for-css-modules-loader
There is a complicated config in preact.config.js, which seems to try
to generate .d.ts files from .css files so typescript can type-check
references to style classes.

This seems to not have been in effect since `typed-css-modules` was
added and run in the npm postinstall script as `npx tcm src
--camelCase dashes`, which also generates all .d.ts files.

postinstall is not the right place for it though, it should run before
every build. It seems to have worked in practice as `make buildweb`
runs `yarn install` as a first step, so this postinstall script is
invoked.