backend/accounts: add running balance and timeseries of transactions
Running balance: balance of the account at each transaction (sum of
all transactions).

Timeseries: balance of account over time.

We add the timeseries of the account balance of the first account to
the account summary chart as a demo. Eventually the account balance
will converted to fiat, so the chart shows the portfolio worth over
time.

Shortcuts to be fixed later:

- Think about how to deal with the local timezone
- Moving the chart will query the backend for the new date range
- Sum of all accounts, not just one.
- A dedicated endpoint for the chart? For a quick demo, I just added
  it to the existing account-summary endpoint.
- `ErrNotAvailable` not handled yet, which is returned if data is
  missing (e.g. block headers not synced, which provide the timestamps
  for the transactions).