backend/coins/eth: remove node/transaction customization options
Main motivation: we have too many NewEtherScan() scattered all over
the place, and use of those depended on various config options and
hardcoded backend values.

To obey the rate limit, we can have at most one EtherScan instance (or
http client/transport inside, but the path to get there is basically
the same).

History:

eth.Coin can could a transaction source (download transactions for an
address) and a ETH standard RPC full node url.

The idea was that if someone wanted to run a full node, they could
disable the trasaction source and get away with showing the balance
only and making transactions. The alternative option was EtherScan.

For the RPC node, we started out with Infura, then moved to
Etherscan. `etherscan+` was a node URL to use EtherScan's RPC proxy
API.

All of this was in the config, and meant to be added to the frontend
for customization.

This turned out to be overkill, and complicates the code considerably.

This commit removes almost all of this complexity, but keeps the
RPCClient and TransactionsSource interfaces, in case it becomes useful
again.