backend/rates: stop all RateUpdater's goroutines when Close'ing backend
This is useful for an upcoming commit where RateUpdater starts using a
persistent database for historical exchange rates which typically needs
to terminate all transactions on shut down.

It is also useful in testing where multiple instances of backends may
introduce lots of loose updater goroutines, if not shut down at the
end of the test.