backend/rates: add jitter to history loops and make them faster
The forward history update loop was iterating every 10min. This
interval was chosen after an approximate time of a new block on the
chain. However, it may be too slow in the worst case where RateUpdater
provides data stale up to 10min while already available from the API.

This commit reduces the potential staleness to only up to ~1.5min.