backend/handlers: show current total independent of chart data
We go back to showing the chart data until the latest date available,
even if it is not completely up to date, so that the user does not
always see 'gathering informatoin' when launching the app.

The total is still computed and shown based on the latest known rates
to match the summary tables.