backend/rates: revert a buggy start/end computation bit in af5da49
af5da49 introduced a new case in computing how long to backfill for
in a loop iteration. Specifically, when fetched data is worth
of less than 90 days.

My empirical testing here shows it may fall into many edge case traps
and be stuck somewhere at 89.x days, never backfilling further in the
past.

It may have already become too complicated and not worth it. I propose
to simplify it back to what it was before: get daily 90days, hourly
anything after. If an API response to the first call falls a bit short
of 90 day, so be it. Forward history updates are hourly anyway.