util/ratelimit: make LimitedCall.Call cancellable by a context
A Call fn could be blocked for too long and had no way to cancel.
For example, this happens when a RateUpdater has many active coin/fiat
pairs and the app settings are changed. The RateUpdater needs to then
quickly reconfigure and shutdown its goroutines but they would be
dangling inside a LimitedCall.Call.

The new context arg in Call allows to get rid of those dangling
goroutines.