backend: make changes to accounts atomic
We had an accountsLock, but it not held for the whole duration of
initAccounts(), which means that in concurrent calls, the accounts
could be uninitialized twice, then all accounts would be added twice,
or other such race conditions.

This commit only acquires lock in exported functions to make it work
with all combinations of account-changing function calls.