frontend: support internal links from guide entries
Currently only external links are supported, opening the url in
the default browser.

This allows to link to internal view, i.e. to '/exchanges'.

The A component now imports and depends on the extConfig function
that is used for API_TOKEN etc. in util/config.js.
The jest setup currently does not support extConfig.