frontend: fix sidebar overlay should cover moonpay iframe
Sidebar overlay that is only visible on smaller screens has
a z-index of 4001, the loading spinner component had one at 4000.

Changing MoonPay's iframe to 4000 would work as its order in the
dom is after the spinner, but could break easier in the future.

Changed Spinner to 2000 and MoonPay to 3000.

FYI other z-index are dialog:4010, guide:4002, sidebar:4001,
wait-dialog:10000 and chart:12.