frontend: add buy bitcoin or buy crypto in moonpay view
Added same check isBitcoin as in info.tsx and fixed a CSS
issue in combination with the spinner component. The spinner
is always in the background and covered later by the MoonPay
iframe as we just let it load and don't have the info if it
really loaded to hide the spinner again. The CSS issue was
that the spinner covered the title.