frontend: allow Moonpay to request camera access
They state that from March 31st, 2021, their in-app KYC process will
require camera access.

They said that
https://webrtc.github.io/samples/src/content/getusermedia/gum/ can be
used to test if it is working. This commit makes the camera at that
site work in the moonpay.tsx iframe.

In the Qt app, the security origin scheme check is dropped, as that is
easier than checking for the moonpay url. Note that all external
requests are blocked anyway, so only already whitelisted pages can
even be loaded, so the check was redundant.