frontend/app: fix bb02 success screen by fixing devices key attribute
After creating a new wallet or restoring from backup, the app is
supposed to show a success screen. This has not happened due to a
regression.

The screen was skipped because the DeviceSwitch and the BitBox02
components were recreated completely, losing the internal state.

The components were recreated because the key attribute,
e.g. `key={['device-switch', devices]}`, "changed". The devices object
was the same, but in Javascript, one can't compare arrays like
this. `[] == []` or `['device-switch'] == [''device-switch']` is
always false, so the component was re-recreated.

We fix it by turning making the key a string derived from the devices,
so preact will correctly identify when the devices changed and when
they didn't change.