frontend/bitbox02: fix sd card inserted state
The BackupsV2 component would be fed the sdCardInserted property via
the URL from the Settings component. This is fragile, e.g. if you
remove the sdcard while in the settings screen and then go to backups,
you will get an error.

The url part is also only used for the BitBox02, but the ManageBackups
component is for both BitBox01 and BitBox02, which makes this also
confusing.

For now we abstract away the microSD card check and use it when
loading the Backups component, so it should always work.

There is more refacoring that can be done later, e.g. cleaning up the
back button mess.