backend: persist default set of accounts for a new keystore
Before, we would load them dynamically whenever a keystore is
connected.

Now, we only load accounts from the accounts.json database, and
pre-populate the database with a default set of accounts to kick
things off.

We take into account the global settings for which coins are active,
so that users who already disabled some accounts have a smooth
migration. The global settings to activate/deactivate coins and tokens
will be removed and replaced with functionality to add and remove
accounts per keystore.