frontend: support non stacked coin icons
The logos of the coins render as stacked icons that is somehow
used to have a hover effect in the sidebar. Technically this
exports 2 versions of the coin logo which is unnecessary when used
outside of the sidebar.

This ads a stacked property the the logos, that is disabeld by
default.