backend: error on duplicate account name
To avoid confusion, we disallow adding an account with a name that
another account already has.

We check all accounts, even ones that are not currently visible - we
might show them in the future in watch-only mode.

We centralize the error codes in the i18n json file so that they can
be reused from multiple endpoints and pages.