backend: check account support by keystore also when loading
We check which accounts a keystore supports when persisting the
default accounts. We should do the same when loading them. The edge
case is that a restores a seed previously used in a BB02 Multi in a
BB02 BTC-only. The app would then show the previously persisted
altcoins even though the connected keystore doesn't support them.

This will come in handy with watch-only, where we have to know which
accounts can be actively used vs. just watched.