backend/aopp: add user-approval state
If an AOPP link is clicked while no BitBox is connected, we do as
before and ask the user to insert and unlock.

However, if the BitBox is already inserted and unlocked, we don't want
a link click to immediately go to the account-chooser, but have an
intro dialog first asking the user if they want to proceed.