frontend: add dialog button component
Currently components have to import dialog styles to style buttons
in a dialog. Also these buttons are currenlty stacked on top of
each other instead of next to each other.

Added a separate DialogButtons component export to reuse the
container to place buttons in a dialog. One single button should
be use 100% width, but 2 buttons should be positioned bottom left and
bottom right of the container.